requirements.txt
csv==1.0
pickleshare==0.7.5 #preguntar si és aquesta la bona, pq l'altre no apareix a la terminal
urllib3==1.26.4
networkx==2.5.1
osmnx==1.1.0
haversine==2.3.0
staticmap==0.5.5
python-telegram-bot==13.5